* CC: do not reset the recorded macro expansion stack when it exceed the limit.
Also, a test case is added.
If we have a macro definition "#define member FF.member", and expanding the usage.
"FF" could also be expanded many times, which exceeds the s_MaxMacroReplaceDepth
When initially expand the "FF.member", the anchor point is recorded as below.
                                    ^-----anchor point
This protect the "member" get expand again. But if we reset the m_ExpandedMacros when
expanding the "FF", then the "member" will expand again, which leads to infinite loop.

This fix bug reported here:
http://forums.codeblocks.org/index.php/topic,20568.0.html

git-svn-id: svn://svn.code.sf.net/p/codeblocks/code/trunk@10499 2a5c6006-c6dd-42ca-98ab-0921f2732cef